+++ site-cookbooks/algieba/attributes/default.rb
@@ -8,3 +8,4 @@ default[:algieba][:shared_dirs] = %w[log tmp bundle]
 default[:algieba][:ruby_version] = '2.4.4'
 default[:algieba][:open_ports] = %w[80:tcp 3000:tcp]
 default[:algieba][:mysql_users] = %w[development test]
+default[:algieba][:requirements] = %w[cmake mysql-devel]