+++ site-cookbooks/denebola/attributes/default.rb
@@ -5,7 +5,7 @@ default[:denebola][:repository] =
 default[:denebola][:branch] = 'develop'
 default[:denebola][:symlinks] = {'backup' => 'backup'}
 default[:denebola][:shared_dirs] = %w[backup/race_list backup/races log bundle]
-default[:denebola][:requirements] = %w[mysql-devel]
+default[:denebola][:requirements] = %w[cmake mysql-devel]
 default[:denebola][:ruby_version] = '2.4.4'
 default[:denebola][:export_dir] = '/etc/exports.d'
 default[:denebola][:exports] = [