+++ site-cookbooks/regulus/recipes/app.rb
@@ -34,10 +34,6 @@ deploy node[:regulus][:deploy_dir] do
       to File.join(node[:regulus][:deploy_dir], 'shared/bundle')
     end
 
-    node[:regulus][:requirements].each do |package_name|
-      package package_name
-    end
-
     execute "#{rvm_do} bundle install --path=vendor/bundle  --clean" do
       cwd release_path
       environment 'PATH' => node[:rvm][:path]
@@ -118,6 +114,10 @@ deploy node[:regulus][:deploy_dir] do
 
     execute "docker rm #{node[:regulus][:app_name]}" do
       only_if "docker ps -a | grep #{node[:regulus][:app_name]}"
+    end
+
+    service 'docker' do
+      action :restart
     end
 
     script_path = File.join(release_path, 'scripts')