+++ site-cookbooks/zosma/recipes/app.rb
@@ -39,10 +39,6 @@ deploy node[:zosma][:deploy_dir] do
       environment 'PATH' => node[:rvm][:path]
     end
 
-    node[:zosma][:requirements].each do |package_name|
-      package package_name
-    end
-
     execute "#{rvm_do} bundle exec rake db:create" do
       cwd release_path
       environment 'PATH' => node[:rvm][:path]